src/requirements.txt
redis
Flask
Flask-Cors
Flask-RESTful
Flask-SQLAlchemy
passlib
cffi
docopt
gunicorn
Flask-Script
python-dateutil
psycopg2-binary
xlrd
numpy
celery
eventlet
requests
sentry-sdk[flask]
riotwatcher